server/requirements.txt
# Web Framework
fastapi
uvicorn[standard]
python-multipart

# LangChain & Ecosystem
langchain
langchain-community
langchain-core
langchain-groq
langchain-google-genai
# Vectorstore
pinecone

# PDF Parsing
pypdf

# Environment Variables
python-dotenv

# Typing & Utilities
pydantic
requests
tqdm

# Logging
loguru